Small refactoring:
- Added cache warmer that gets the tree from ContentStoreService to cache merged/processed items.
- Fixed DeploymentEventsWatcher issues.
