+++ src/main/java/org/craftercms/engine/cache/SiteCacheWarmer.java
@@ -14,24 +14,25 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.craftercms.engine.util.cache;
+package org.craftercms.engine.cache;
 
-import org.craftercms.core.service.Context;
+import org.craftercms.engine.service.context.SiteContext;
 
 /**
- * Helper class that's used to warm up the cache of a context. Normally several of these will be used by the
- * {@link SiteCacheWarmer} to do a full warm up.
+ * Helper class that's used to warm up the cache of a site context.
  *
  * @author avasquez
  * @since 3.1.4
  */
-public interface ContextCacheWarmer {
+public interface SiteCacheWarmer {
 
     /**
-     * Warm ups the cache associate to the context
+     * Warms up the cache for the specified site context. If {@code switchCache} is true, then a new cache is created,
+     * warmed up and then switched with the current cache of the site.
      *
-     * @param context the context
+     * @param siteContext the site context
+     * @param switchCache if a new cache should be warmed and then switched with the current cache
      */
-    void warmUpCache(Context context);
+    void warmUpCache(SiteContext siteContext, boolean switchCache);
 
 }
