+++ src/main/resources/crafter/engine/server-config.properties
@@ -257,7 +257,7 @@ crafter.engine.graphql.query.sort.order=DESC
 crafter.engine.site.cache.warmUp.enabled=false
 # The descriptor folders that need to be preloaded in cache, separated by comma. Specify the preload depth with
 # :{depth} after the path. If no depth is specified, the folders will be fully preloaded.
-crafter.engine.site.cache.warmUp.descriptor.folders=/site:5
+crafter.engine.site.cache.warmUp.descriptor.folders=/site:3
 # The content folders that need to be preloaded in cache, separated by comma. Specify the preload depth with
 # :{depth} after the path. If no depth is specified, the folders will be fully preloaded.
 crafter.engine.site.cache.warmUp.content.folders=/scripts,/templates
